package/ecl: Initial commit of ecl project
Uses ecl release 16.1.3 with a stock ARM cross_config from the
20.4.24 release. Required for ts7820-utils package.

Bumping to 20.4.24 is possible, however, the static library used
in the ts7820-utils repo is built using 16.1.3 so that needs to be
built here.

Note that this version of ecl seems to require BR2_PACKAGE_BDWGC
as there is an issue with Buildroot and ecl's builtin libgc. When
moving to 20.4.24 (or other future releases) this library may be
able to be removed from the config.